requirements.txt
newspaper3k==0.1.9
peewee==2.9.1
psycopg2==2.7.1
pytest==3.0.7
pytest-mock==1.6.0
pytest-randomly==1.1.2
pyyaml==3.12.0
